+++ ui/js/component/fileCardStream/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import {
+  connect
+} from 'react-redux'
+import FileCardStream from './view'
+
+export default connect()(FileCardStream)
