+++ ui/js/component/fileTile/index.js
@@ -2,9 +2,13 @@ import React from 'react'
 import {
   connect
 } from 'react-redux'
+import {
+  selectResolvedUris,
+} from 'selectors/content'
 import FileTile from './view'
 
 const select = (state) => ({
+  resolvedUris: selectResolvedUris(state),
 })
 
 const perform = (dispatch) => ({
